[physics] Deprecate GeometryMovement
Move to _fluid_legacy.py
