+++ phi/physics/_fluid_legacy.py
@@ -2,8 +2,8 @@ import warnings
 
 from phi import math, struct, field
 from phi.field import CenteredGrid, StaggeredGrid
-from . import advect
-from ._effect import Gravity, effect_applied, gravity_tensor
+from . import advect, Obstacle
+from ._effect import Gravity, effect_applied, gravity_tensor, FieldEffect
 from ._physics import Physics, StateDependency, State
 from .fluid import make_incompressible
 
@@ -128,3 +128,22 @@ class IncompressibleFlow(Physics):
             'divergent_velocity': divergent_velocity,
         }
         return fluid.copied_with(density=density, velocity=velocity, age=fluid.age + dt, solve_info=solve_info)
+
+
+class GeometryMovement(Physics):
+
+    def __init__(self, geometry_function):
+        Physics.__init__(self)
+        self.geometry_at = geometry_function
+
+    def step(self, obj, dt=1.0, **dependent_states):
+        next_geometry = self.geometry_at(obj.age + dt)
+        h = 1e-2 * dt if dt > 0 else 1e-2
+        perturbed_geometry = self.geometry_at(obj.age + dt + h)
+        velocity = (perturbed_geometry.center - next_geometry.center) / h
+        if isinstance(obj, Obstacle):
+            return obj.copied_with(geometry=next_geometry, velocity=velocity, age=obj.age + dt)
+        if isinstance(obj, FieldEffect):
+            with struct.ALL_ITEMS:
+                next_field = struct.map(lambda x: x.copied_with(geometries=next_geometry) if isinstance(x, GeometryMask) else x, obj.field, leaf_condition=lambda x: isinstance(x, GeometryMask))
+            return obj.copied_with(field=next_field, age=obj.age + dt)
